IPC host and port handling reimplemented
IPC server takes bind port as argument to Start method so it is able to check if can bind
socket directly or not and throw exception if port in use by other process.

When port 0 is provided, IPC server selects random free port to bind to.

IPC client requests its dispatcher for host:port combination before each attempt to connect
to IPC server.